fix: resolve pricing page 500 by separating CREDIT_PACKS from Stripe SDK
The pricing page (Server Component) imported CREDIT_PACKS from
stripe/config.ts, which also imports the stripe Node.js SDK. This
caused a 500 error because the stripe package fails to bundle correctly
in Next.js 16 Server Components.

Fix: extract CREDIT_PACKS to its own module (credit-packs.ts) with no
stripe dependency. The pricing page now imports from credit-packs.ts
directly. Existing imports from config.ts continue to work via re-export.
