+++ src/lib/stripe/config.ts
@@ -1,5 +1,8 @@
 import Stripe from "stripe";
 
+// Re-export credit packs for backward compatibility
+export { CREDIT_PACKS } from "./credit-packs";
+
 let _stripe: Stripe | null = null;
 
 export function getStripe() {
@@ -11,30 +14,3 @@ export function getStripe() {
 
 // Keep backward compat
 export const stripe = null as unknown as Stripe;
-
-export const CREDIT_PACKS = {
-  starter: {
-    name: "Starter",
-    credits: 3,
-    price: 4.99,
-    pricePerSong: 1.66,
-    priceId: process.env.STRIPE_STARTER_PACK_PRICE_ID || "",
-    popular: false,
-  },
-  popular: {
-    name: "Family Pack",
-    credits: 10,
-    price: 9.99,
-    pricePerSong: 1.0,
-    priceId: process.env.STRIPE_FAMILY_PACK_PRICE_ID || "",
-    popular: true,
-  },
-  mega: {
-    name: "Mega Pack",
-    credits: 25,
-    price: 19.99,
-    pricePerSong: 0.8,
-    priceId: process.env.STRIPE_MEGA_PACK_PRICE_ID || "",
-    popular: false,
-  },
-} as const;
